Read initial state from data when fitting
When fitting a protocol where the control variable's initial state is
far from nominal and the first segment is a rate-controlled change in
the control variable, it is necessary to update the protocol's initial
state to match the data or the estimated segment durations will be very
wrong.  Since the fit is penalized by the difference in fitted and
nominal duration, the resulting fit will also be wrong.

Therefore, update the protocol's initial state to match the data before
fitting it to the data.