Rework state calculations for free variables
Default values for variables pose a problem when calculating state from
a Segement object because the Segment may not be aware of all the
variables that need default values.  I initially tried to solve this
problem by make Segment's constructor take a list of variables as well
as a list of transitions, the idea being that any variables that were
missing transitions should be interpreted according to the defaults
specified in the Protocol.  This solution isn't reliable though because
Segment objects can be constructed or modified outside protocols.
Therefore Segment.eval_state() must take a list of variables; variables
in this list that are not otherwise constrained should be interpreted as
free.

This commit also contains a placeholder hack to handle transitions
applied to previously unconstrained variables, which bring them into a
relatively-constrained state.  Without support for values referenced to
past variable values, this can't be handled properly, so for now the
state calculations consider an unconstrained (free) variable to remain
unconstrained.